backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
Pillow==11.0.0
opencv-python==4.8.0.76
numpy==1.24.3
scipy==1.11.4
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
requests==2.32.3
python-dateutil==2.9.0
h5py==3.10.0
Werkzeug==3.1.3
